Get proposal information on frontend
